course redirected to admin course view page
